OilPvtThermal: move implementation to impl file
